+++ appconfig.go
@@ -3,15 +3,16 @@ package configuration
 import (
 	"github.com/containerssh/auditlog"
 	"github.com/containerssh/auth"
-	"github.com/containerssh/dockerrun"
+	"github.com/containerssh/docker"
 	"github.com/containerssh/geoip"
-	"github.com/containerssh/kuberun"
+	"github.com/containerssh/kubernetes"
 	"github.com/containerssh/log"
 	"github.com/containerssh/metrics"
 	"github.com/containerssh/sshserver"
 )
 
 // AppConfig is the root configuration object of ContainerSSH.
+//goland:noinspection GoDeprecation
 type AppConfig struct {
 	// Listen is an alias for ssh.listen. Its usage is deprecated.
 	Listen string `json:"listen,omitempty" yaml:"listen,omitempty" default:""`
@@ -39,20 +40,25 @@ type AppConfig struct {
 
 	// Backend defines which backend to use. This option can be changed from the config server.
 	Backend string `json:"backend" yaml:"backend" default:"dockerrun" comment:"Backend module to use"`
-	// DockerRun contains the configuration for the dockerrun backend.
-	DockerRun dockerrun.Config `json:"dockerrun" yaml:"dockerrun" comment:"Docker configuration to use when the Docker run backend is used."`
-	// KubeRun contains the configuration for the kuberun backend.
-	KubeRun kuberun.Config `json:"kuberun" yaml:"kuberun" comment:"Kubernetes configuration to use when the Kubernetes run backend is used."`
+	// Docker contains the configuration for the docker backend.
+	Docker docker.Config `json:"docker,omitempty" yaml:"docker" comment:"Docker configuration to use when the Docker backend is used."`
+	// DockerRun contains the configuration for the deprecated dockerrun backend.
+	DockerRun docker.DockerRunConfig `json:"dockerrun,omitempty" yaml:"dockerrun" comment:"Docker configuration to use when the Docker run backend is used."`
+	// Kubernetes contains the configuration for the kubernetes backend.
+	Kubernetes kubernetes.Config `json:"kubernetes,omitempty" yaml:"kubernetes" comment:"Kubernetes configuration to use when the kubernetes run backend is used."`
+	// KubeRun contains the configuration for the deprecated kuberun backend.
+	KubeRun kubernetes.KubeRunConfig `json:"kuberun,omitempty" yaml:"kuberun" comment:"Kubernetes configuration to use when the kuberun run backend is used."`
 }
 
+// FixCompatibility moves deprecated options to their new places and issues warnings.
 func (cfg *AppConfig) FixCompatibility(logger log.Logger) error {
 	if cfg.Listen != "" {
 		if cfg.SSH.Listen == "" || cfg.SSH.Listen == "0.0.0.0:2222" {
-			logger.Warningf("you are using the deprecated 'listen' option for SSH listen socket instead of the new 'ssh -> listen', please change your configuration")
+			logger.Warningf("You are using the 'listen' option deprecated in ContainerSSH 0.4. Please use the new 'ssh -> listen' option. See https://containerssh.io/deprecations/listen for details.")
 			cfg.SSH.Listen = cfg.Listen
 			cfg.Listen = ""
 		} else {
-			logger.Warningf("you are using both the deprecated 'listen' and the new 'ssh -> listen' options, the new option takes precedence")
+			logger.Warningf("You are using the 'listen' option deprecated in ContainerSSH 0.4 as well as the new 'ssh -> listen' option. The new option takes precedence. Please see https://containerssh.io/deprecations/listen for details.")
 		}
 	}
 	return nil