+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.9.4: New docker and kubernetes backends
+
+This release adds support for the new [docker](https://github.com/containerssh/docker) and [kubernetes](https://github.com/containerssh/kubernetes) backends.
+
 ## 0.9.3: Metrics integration
 
 This release integrates the [metrics library](https://github.com/containerssh/metrics) which is now required as a dependency when creating a HTTP client.